break app into several handlers
