+++ index.js
@@ -8,13 +8,12 @@ const buildSchemas = require('./src/schemas');
 
 const seed = require('./src/seed');
 
-const appConfig = require('./src/app');
+const app = require('./src/app')();
 
 const logger = require('./src/logger.js');
 
 (async () => {
   await buildSchemas(db);
   seed(db);
-  const app = appConfig(db);
   app.listen(port, () => logger.info(`App started and listening on port ${port}`));
 })();
