+++ src/handlers/post-ride.js
@@ -1,0 +1,59 @@
+const db = require('../db');
+
+module.exports = async (req, res) => {
+  const startLatitude = Number(req.body.start_lat);
+  const startLongitude = Number(req.body.start_long);
+  const endLatitude = Number(req.body.end_lat);
+  const endLongitude = Number(req.body.end_long);
+  const riderName = req.body.rider_name;
+  const driverName = req.body.driver_name;
+  const driverVehicle = req.body.driver_vehicle;
+
+  if (startLatitude < -90 || startLatitude > 90 || startLongitude < -180 || startLongitude > 180) {
+    return res.status(400).send({
+      error_code: 'VALIDATION_ERROR',
+      message: 'Start latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively'
+    });
+  }
+
+  if (endLatitude < -90 || endLatitude > 90 || endLongitude < -180 || endLongitude > 180) {
+    return res.status(400).send({
+      error_code: 'VALIDATION_ERROR',
+      message: 'End latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively'
+    });
+  }
+
+  if (typeof riderName !== 'string' || riderName.length < 1) {
+    return res.status(400).send({
+      error_code: 'VALIDATION_ERROR',
+      message: 'Rider name must be a non empty string'
+    });
+  }
+
+  if (typeof driverName !== 'string' || driverName.length < 1) {
+    return res.status(400).send({
+      error_code: 'VALIDATION_ERROR',
+      message: 'Driver name must be a non empty string'
+    });
+  }
+
+  if (typeof driverVehicle !== 'string' || driverVehicle.length < 1) {
+    return res.status(400).send({
+      error_code: 'VALIDATION_ERROR',
+      message: 'Driver vehicle must be a non empty string'
+    });
+  }
+
+  var values = [req.body.start_lat, req.body.start_long, req.body.end_lat, req.body.end_long, req.body.rider_name, req.body.driver_name, req.body.driver_vehicle];
+
+  try {
+    const { statement } = await db.execute('INSERT INTO Rides(startLat, startLong, endLat, endLong, riderName, driverName, driverVehicle) VALUES (?, ?, ?, ?, ?, ?, ?)', values);
+    const { rows } = await db.query('SELECT * FROM Rides WHERE rideID = ?', statement.lastID);
+    res.send(rows);
+  } catch (err) {
+    return res.status(500).send({
+      error_code: 'SERVER_ERROR',
+      message: 'Unknown error'
+    });
+  }
+};
